Expose endpoint to get list of UA activable services
The list of activable UA services was only retrieved from the
client-side of Subiquity (using ubuntu-advantage-tools). Therefore, the
desktop installer would need to reimplement the same logic should they
need access to the list of services ; which is inconvenient.

We now expose a new endpoint in the API that takes the token as
a parameter and returns a status (+ a list of services if the token is
valid and not expired).

  $ curl \
    --unix-socket .subiquity/socket \
    --header 'Content-Type: application/json' \
    http://a/ubuntu_advantage/check_token \
    -d '"C123456"'

The token parameter is expected to be in the body of the request -
rather than in the query string - to avoid ending up in the access logs.

This new endpoint is a read-only GET endpoint. It is not designed as a
replacement for the POST to /a/ubuntu_advantage that the client must
(still) do to pass the token to the model.

We now use this new endpoint internally in Subiquity so that the
retrieval of the activable services is done on the server-side only.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>