Fix hanging when network is the last postinst model configured
When running an autoinstallation with only “network” as an interactive
section, the installation hanged after the following steps:

  finalizing installation
    running 'curtin hook'
      curtin command hook
  executing late commands

where in normal conditions, the next step is:

  final system configuration

Also, querying the meta/status endpoint at this point shows that the app
is in POST_WAIT state.

The problem is:

the network model is declared as both an “install” model and a
“postinstall” model (it is the only one we have AFAIK)
when calling .configured() for a given model, we only trigger /at max/
one event. Either:
 * install_event (if the model is an “install” model - and it is the
   last one configured)
 * postinstall_event (if the model is a “postinstall” model but not an
   “install” model - and it is the last one configured)
 * no event (as in most cases)

So when we call .configured() for the network model, we /can/ trigger
the install_event but can never trigger the postinstall event.
Therefore Subiquity, will wait forever until something triggers the
postinstall_event

Fixed by accepting to trigger the install_event and the
postinstall_event in a single call to .configured.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>