Add script to validate autoinstall data against JSON schema
The script can be used to validate autoinstall user data against the
schema. By default, it expects a #cloud-config header and the user-data
to be under the autoinstall: key.

By passing the --no-expect-cloudconfig, it validates the data directly.
We can use this option to validate the YAML files under
examples/autoinstall-*.yaml

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>