Make sure SSHController always yields valid autoinstall data
When running an autoinstallation with no ssh: field (which is valid),
the output autoinstall-user-data file would end up with None for the
key ssh:authorized-keys, as shown below:

  ssh:
    authorized-keys: null

Unfortunately, null is not a valid value according to the schema (which
expects a string).

Fixed by initializing the authorized_keys variable to [] instead of None
so that it yields an empty list instead of null when no autoinstall data
is loaded.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>