Add option to run commands with private mounts
The LoggedCommandRunner now has the option to start a command with
private mounts enabled - see mount_namespaces(7).

To use this feature, we now rely on systemd-run instead of systemd-cat.
The main benefit is that it should allow us not to worry about mounts
refusing to be un-mounted after executing a curtin in-target command.

On the downside, systemd-run requires root privileges (unless --user is
available) so we can't use it in dry-run mode. For dry-run mode, we
still need to rely on systemd-cat.

Another downside is that the commands passed to systemd-run are run in a
clean environment. Therefore, we must explicitly set the environment
variables that we want to pass to the commands to execute.

I came up with a short list of environment variables that we need but we
might need to add more later.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>