filesystem/actions: fix action query getting None
These gap check loops were falling out of the end of the function, thus
returning None instead of False.  This resulted in the Action handler
having to cope with a None, and the caller trying to unpack a None.

Add a default string for if this happens again, attempt to coax other
values into a displayable reason.