Implement backward compatibility mechanism for autoinstall key
In addition to autoinstall_key, controllers can now define an
autoinstall_key_alias.

The purpose of this attribute is to provide a backward compatibility
mechanism when we need to change the name of an autoinstall_key for
whatever reason.

When loading autoinstall data, if we can't find data matching the
autoinstall_key, we will now also look for data matching the key alias.

In the JSON schema, the old name (i.e., autoinstall_key_alias) will be
marked deprecated and the description will have the name of the
replacement key (i.e., autoinstall_key).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>