autoinstall: specific order of possible locations
To allow customization where an install iso can have the autoinstall
data, and to ignore that in favor of autoinstall from cloud-config, this
implements a more specific order of loading from the possible locaitons.

Also, the chosen file is set aside, and then is top priority if the
subiquity process is reloaded.