tests: rename mock variable arun_command
Having a variable named arun_command in tests interferes with tags
generated with ctags: the default tag does not point to the actual
function definition; but to the variable in test_ubuntu_advantage.py.
Excluding the tests when generating the tags would be an option but for
now we'll just rename the variable.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>