storage: fix screen sometimes not refreshing after slow probing
When we reach the storage screens on the installer, if the devices
probing operation has not finished, we:
 * display a temporary "Probing" screen.
 * create an asynchronous task (a.k.a., probing task) that will
   eventually show the "Guide Storage" screen when the probing operation
   finishes.

The probing task checks, when it finishes, that the screen currently
visible is the "Probing" screen. This is the expectation and is true in
most scenarios. But in case a different screen is visible, we skip
refreshing the display.

Unfortunately, sometimes, a "Progress" screen is shown for some time
before the "Probing" screen appears. Consequently, we do not refresh the
screen if the probing operation finishes whilst the Progress screen is
visible.

In order to keep the view returned by make_ui() up-to-date and make sure
that the right screen is shown even if the probing operation finishes
early, we use the level indirection that was implemented in make_ui.

https://bugs.launchpad.net/subiquity/+bug/1968161

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>