source: add support for autoinstall data
Since we added a search_drivers checkbox that is uncheckd by default,
there is no longer a way for users to install third-party drivers in an
autoinstall context.

We now implement the autoinstall support for source so that users can
specify what value they want for search_drivers.

Futhermore, to be backward compatible with existing autoinstall
configurations, we now make search_drivers default to true in
autoinstall contexts.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>