ubuntu-pro: move mode selection in a different screen
We now split the Ubuntu Pro screen into two sub views.
The first view asks whether the user wants to skip or upgrade (we call
this the upgrade_yes_no_screen in the code).

The second one asks the user to input the contract token. Going forward,
this view will also be used for Ubuntu One.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>