ubuntu-pro: don't assume an overlay is present in show_* functions
show_* functions from the UbuntuProView object always assumed that the
loading overlay was present. When called, they would therefore attempt
to close the overlay. This results in a crash if no overlay is present.

With the current design, it is fine because we never call the show_*
functions without an overlay being present.

However, with new design of Ubuntu Pro, sometimes the show_* functions
will be called without an overlay.

We now delegate the responsibility of removing the overlay by using a
callback system.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>