autoinstall: reject null values in autoinstall top level sections
According to the autoinstall schema, the top level autoinstall sections
(except proxy) should not accept null as a valid value.

However, passing null in one of these section effectively bypasses the
schema validation.

The reason why we decided to skip validation when a section is null is
to allow the autoinstall_default key of the controllers to be None.
Having None as a default value is a perfectly sensible choice, but the
logic is sort of wrong.

Instead of skipping the validation when a section is resolved to None,
we should instead skip validation if the section was not present.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>