ubuntu-pro: fix missing return statement
The function is_token_added_overlay supposedly returns a boolean. But
the absence of a final return False statement made it return None by
accident.

In practice, it should not make a difference because returning False or
None evaluate to False in a boolean context. But returning False is
cleaner (sorry Perl).

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>