cmdlist: refactorize early/late/error commands
A new class called Command is introduced. A Command instance has
arguments (i.e., either a string or an sequence of strings) and a
boolean called "check" that tells if an error should stop the
installation.

A Command instance also has helper methods to simplify its usage.

The early/late/error command controllers now have a sequence of Command
instances.

When loaded from an autoinstall config, the commands inherit the value
of the "check" attribute from the controller itself.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>