snaplist: fix crash when pressing "Try again" on snaplist
When pressing the "Try again" button on the snaps list, Subiquity would
crash with:

TypeError: SnapListView.wait_load() takes 1 positional argument but 2
were given

This happens because urwid passes the sender to the callback - which we
need to discard or accept it in the signature.

Fixed by discarding the sender.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>