snaplist: avoid showing snaplist screen on failure
When creating the UI for snaplist, if the server returns a FAILED status
when we request the list of snaps, we want to move on to the next
screen.

However, calling self.next_screen in the make_ui function while also
returning an actual view triggers a race condition. Depending on the
order of execution of the instructions (which are asynchronous), the
snaplist view can end up being shown even though we moved to the next
screen. This ends up doing funny things upon clicking done.

Fixed by raising the Skip exception after making sure we mark snaplist
configured.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>