mirror: only start configuring APT once source & mirror are configured
The mirror controller is responsible from creating the APT configuration
once its associated model is marked configured. However, the APT
configuration is also dependent on the source model being marked
configured.

When subiquity is used for the server installer, the client immediately
sends a POST /a/meta/client_variant request. This ends up marking the
source model as configured (even though the default value does not make
much sense). So we never end up in a scenario where the mirror model is
marked configured before the source model.

When Subiquity is used by a different client, such as the
ubuntu-desktop-installer, it is possible that we mark the mirror model
configured first. This ends up crashing the installer.

Fixed by adding a dependency on the source model so that we do not end
up creating the APT configuration too early.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>