snaplist: make list snap task raise on error
The list snaps task used to return successfully even in case of error.

To make the distinction between and error and a successful fetch, we
used to rely on the task to set an attribute at the loader level:
 * failed = True; or
 * failed = False

The task was responsible for setting this attribute and there were
scenarios where it would not set the attribute properly. Usually, this
was when an exception was raised in the task.

We now raise an exception when an error occurs in the task ; instead of
returning. This is better because we can know if the task:
 * got cancelled - by calling "task.cancelled()"
 * finished successfully - by calling "task.done() and not
   task.exception()"
 * failed - by calling "task.done() and task.exception()"
 * has not finished - by calling "not task.done()"

This also allows us to detect errors and retry

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>