filesystem: accept extended partition with size -1 in autoinstall
Our documentation of autoinstall mentions that one can specify the size
of a partition as -1 to use the remaining space on the disk.

Because it only makes sense in practice to support this option for the
right-most partition of a given disk, specifying size: -1 is only
accepted when defining the last partition of the disk.

Having said that, when making an autoinstall configuration that includes
an extended partition, the user must define the logical partitions after
the extended partition.

For the following configuration:

|                                  disk                                |
+--------------+--------------+--------------+-------------------------|
| p1 (primary) | p2 (primary) | p3 (primary) | p4 (extended)           |
|              |              |              | p5 (logic) | p6 (logic) |

, the representation in Subiquity will be analogous to:

[
  Partition(p1, "primary"),
  Partition(p2, "primary"),
  Partition(p3, "primary"),
  Partition(p4, "extended"),
  Partition(p5, "logical"),
  Partition(p6, "logical"),
]

This means that specifying size: -1 on p4 gets rejected because it is
not the last partition defined.p6 is the last partition defined.

This is however a valid use-case because p5 and p6 are contained within
p4. So p4 should be able to use the remaining space and there is no
reason to reject this configuration

Fixed by ignoring logical partitions when checking if a size of -1 is
allowed on an extended partition.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>