filesystem: mark ptable msdos in tests at the right place
In the new tests recently added to test_filesystem.py, the initial disk
object was marked having a msdos partition table.

This is fine for computing offsets manually. However, what really needs
to be done is to mark the partition table as msdos in the YAML itself.
Otherwise, the disk is reformatted as GPT.

This makes gap functions use the right alignment data when computing the
list of gaps available.

Sadly it breaks the tests because of other issues, so I added FIXME tags
in the broken tests.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>