filesystem: fix use of negative size in gaps calculations
When size: -1 is specified on a given partition, we attempt to determine
the its right size by looping through the partitions and computing a
list of available gaps.

That said, when looping through the partitions, we also include the one
which has an incomplete (i.e. = -1) size. This makes the computation
rely on a negative value and this often leads to incorrect result.

Fixed by excluding the current partition when determining its
appropriate size.

https://bugs.launchpad.net/ubuntu/+source/subiquity/+bug/1991929

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>